marked 1 test where API key is needed ruff formatting
