+++ tests/applications/cli/test_collect.py
@@ -2,7 +2,6 @@
 Tests the collect_learnings function in the cli/collect module.
 """
 
-
 import pytest
 
 # def test_collect_learnings(monkeypatch):
