+++ tests/core/default/test_simple_agent.py
@@ -7,12 +7,21 @@ from gpt_engineer.core.default.paths import ENTRYPOINT_FILE
 from gpt_engineer.core.default.simple_agent import SimpleAgent
 from gpt_engineer.core.files_dict import FilesDict
 from gpt_engineer.core.prompt import Prompt
+from tests.mock_ai import MockAI
+from langchain.schema import AIMessage
 
 
 def test_init():
     temp_dir = tempfile.mkdtemp()
-
-    lean_agent = SimpleAgent.with_default_config(temp_dir, CachingAI())
+    mock_ai = MockAI(
+        [
+            AIMessage(
+                f"hello_world.py\n```\nwith open('output.txt', 'w') as file:\n    file.write('Hello World!')\n```"
+            ),
+            AIMessage(f"```run.sh\npython3 hello_world.py\n```"),
+        ],
+    )
+    lean_agent = SimpleAgent.with_default_config(temp_dir, mock_ai)
     outfile = "output.txt"
     code = lean_agent.init(
         Prompt(
@@ -37,7 +46,14 @@ def test_improve():
             "run.sh": "python3 main.py\n",
         }
     )
-    lean_agent = SimpleAgent.with_default_config(temp_dir, CachingAI())
+    mock_ai = MockAI(
+        [
+            AIMessage(
+                f"```diff\n--- main.py\n+++ main.py\n@@ -7,3 +7,3 @@\n     with open(filename, 'w') as file:\n-        file.write('Hello World!')\n+        file.write('!dlroW olleH')\n```"
+            )
+        ]
+    )
+    lean_agent = SimpleAgent.with_default_config(temp_dir, mock_ai)
     code = lean_agent.improve(
         code,
         Prompt(
