+++ tests/mock_ai.py
@@ -1,0 +1,14 @@
+from typing import List, Any, Optional
+
+
+class MockAI:
+    def __init__(self, response: List):
+        self.responses = iter(response)
+
+    def start(self, system: str, user: Any, *, step_name: str) -> List[str]:
+        return [next(self.responses)]
+
+    def next(
+        self, messages: List[str], prompt: Optional[str] = None, *, step_name: str
+    ) -> List[str]:
+        return [next(self.responses)]
