+++ tests/test_install.py
@@ -1,6 +1,7 @@
 """
 Tests for successful installation of the package.
 """
+
 import shutil
 import subprocess
 import sys
@@ -84,6 +85,7 @@ def test_cli_execution():
     ), f"gpt-engineer command failed with message: {result.stderr}"
 
 
+@pytest.mark.requires_key
 def test_installed_main_execution(tmp_path, monkeypatch):
     # Ignore git installation check
     monkeypatch.setattr("gpt_engineer.core.git.is_git_installed", lambda: False)
